LayoutTests/platform/ios-sim-deprecated/fast/frames/sandboxed-iframe-attribute-parsing-09-expected.txt

CONSOLE MESSAGE: line 9: Error while parsing the 'sandbox' attribute: 'allow_scripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 4: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Misspelling.

PASS
